refactor: CategoryController의 API V1을 NoticeController로 이동
